refactor: move node utilities to gui
