+++ gui/utils/node_utils.py
